pyproject.toml
[tool.poetry]
name = "pdx"
version = "0.1.0"
description = "PDX is a tool-kit for managing and writing prompts for LLMs."
keywords = ["prompt", "LLM", "prompt engineering"]
authors = ["Adithya Krishnan <krishsandeep@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"
packages = [{include = "pdx", from = "src"}]

[tool.poetry.scripts]
pdx = "pdx.cli:main"

[tool.poetry.dependencies]
python = ">=3.7"
click = "^8.1.3"
jinja2 = "^3.1.2"
pyyaml = "^6.0"
jsonschema = "^4.17.3"
pydantic = "^1.10.9"
regex = "^2023.6.3"
aiohttp = "^3.8.4"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
pytest = "^7.3.2"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
